split up gocd module and up to v4
